% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyou_data2019.R
\docType{data}
\name{jikou}
\alias{jikou}
\title{'Zikou' list of Manyoshu data}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1166 rows and 1 columns.
}
\usage{
jikou
}
\description{
'Zikou' list of Manyoshu data
}
\seealso{
Other manyou_data: 
\code{\link{kajin}},
\code{\link{manyou}}
}
\concept{manyou_data}
\keyword{datasets}
