% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyou-package.R, R/manyou_data2019.R
\docType{package}
\name{manyou}
\alias{manyou}
\alias{manyou-package}
\title{manyou: Text Data of the Manyoshu}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4565 rows and 9 columns.
}
\usage{
manyou
}
\description{
Text data of the Manyoshu editted by Yoshimura Makoto and his contributors.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paithiov909/manyou}
  \item Report bugs at \url{https://github.com/paithiov909/manyou/issues}
}


Other manyou_data: 
\code{\link{jikou}},
\code{\link{kajin}}
}
\author{
\strong{Maintainer}: Akiru Kato \email{paithiov909@gmail.com}

Authors:
\itemize{
  \item Makoto Yoshimura \email{y_makoto@yamaguchi-u.ac.jp}
}

}
\concept{manyou_data}
\keyword{datasets}
\keyword{internal}
