% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyou_data2019.R
\docType{data}
\name{wakati}
\alias{wakati}
\title{Pre-tokenized text from 'manyou'}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4562 rows and 2 columns.
}
\usage{
wakati
}
\description{
Pre-tokenized text from 'manyou'
}
\keyword{datasets}
